Move test_percentage_change_gdp from test_macro_elasiticy.py to test_pufcsv.py
